pyproject.toml
[tool.poetry]
name = "outro-teste"
version = "0.1.0"
description = ""
authors = ["ericyaang <ericyang.seeed@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.2.0"
xgboost = "^1.7.4"
pandas = "^1.5.3"
workalendar = "^17.0.0"
datetime = "^5.0"
mlflow = "^2.1.1"
joblib = "^1.2.0"
hydra-core = "^1.3.1"
lightgbm = "^3.3.5"
hyperopt = "^0.2.7"
statsmodels = "^0.13.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
pytest = "^7.2.1"
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
